third_party/blink/web_tests/fast/input/input-device-constructor-expected.txt

This tests the constructor for the InputDeviceCapabilities DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new InputDeviceCapabilities().firesTouchEvents is false
PASS new InputDeviceCapabilities({ firesTouchEvents: true }).firesTouchEvents is true
PASS successfullyParsed is true

TEST COMPLETE
